requirements.txt
# Wirepas Oy 2018
wirepas_messaging==1.1.0

# grpc & protobuf related
grpcio==1.14.1
grpcio-tools==1.14.1

# mqtt
paho-mqtt>=1.3.1

# dbs
mysqlclient>=1.3.13
influxdb>=5.2.1

# data viz
pandas==0.23.4

# mutliprocess
loky==2.4.1

# websocket
websocket-client==0.48.0

# tools
PyYAML>=3.13
fluent-logger==0.9.3

Sphinx>=1.7.0; platform_machine == "x86_64"
sphinxcontrib-napoleon>=0.6.1; platform_machine == "x86_64"
